Remove the autowrap package as an optional dependency
It was never actually a dependency. It was added by
161964db4e37e5cee54c97a7cb7a00a7b84efc75. Presumably the author of that commit
confused the autowrap *package* with the autowrap *SymPy* function/submodule.
But the two are unrelated. A grep of "autowrap" in the codebase reveals that
all references to it refer to the SymPy function, not the external package.

Thank you to Samuel Wein for pointing this out to me via private
correspondence.