Run mypy on a separate build on GitHub Actions
This makes it so that the other tests still run when mypy fails. Although mypy
is a "code quality" check, the code-quality checks on CI cause other builds to
not run, so they should only be things that are so basic that other tests need
not run until they are fixed. Typing errors are often complicated to fix so
they don't generally fall into this category.

Splitting it like this also makes it easy to make it not required to pass if a
mypy update introduces new errors to every pull request, until those are
fixed.