tensor/indexed: Indexed now sympifies base argument
`Indexed` can currently be used to symbolically indicate elements
of a `MutableArray`. However, when doing so that MutableArray is
stored in `Indexed`'s `args`, which should not be allowed since
a `MutableArray` is not a subclass of `Basic`. `Indexed` now uses
`_sympify` on its `base` argument to convert any `MutableArray`
to an `ImmutableArray`.