Test that the stacklevel is set correctly in warns()
The stacklevel makes it so that the warning message shows the actual user line
of code (the default stacklevel=1 just shows the code that called warn()).
This is somewhat ambiguous in a lot of cases, which is why there's a flag to
disable it. The issue is that if the function that issues the warning is not
always the top-level function called by the user, or always in the same call
stack, there's no single value for stacklevel that will show the user code.
However, at worst, a wrong stacklevel will just show a useless line of code,
which is no different from the default stacklevel.

It's mostly useful for deprecation warnings (warns_deprecated_sympy()), since
in that case, we only care about the user calling the deprecated function and
that's it (no other function should call deprecated code).