Update hyperexpand.py
Since tests did not fail, the k being rounded must be equivalent to an integer or else it wouldn't have been removed successfully. So for now, assume that Float entities are integers.

The handling of b from bm at line 2334ff, however, is questionable -- gamma(float) would just give a float and this seems reasonable, so I'm not sure why the conversion to int is happening -- perhaps so the expansion below will work better?