Skip the stacklevel test for some deprecations that involve the cache
This fail when the cache is turned off and they also fail in pypy because pypy
uses a pure Python implementation of functools.lru_cache.

The cache kind of breaks these deprecations anyway (the deprecation can only
be triggered exactly once before it is "cached away"). But I'm not really in
the mood to try to figure out how to make the cache ignore things that trigger
warnings.