Reduce "yield from" overhead in iterables.variations()
Same treatment as was recently applied to iterables.subsets().  All
paths through iterables.variations() still return a generator object,
but variations() is no longer itself a generator function.  This
avoids the overhead of one level of yield from.  This change can save
as much as 30% of the time spent in this function.  (Very sensitive to
Python version.)