Adjust statement about namespace implementation detail.
I came across this statement in the docstring guide and do not consider
module layout an "implementation detail". Modules without a prepended
underscore in their filename should be considered public API because
people will import from them, regardless if a sentence in the
documentation warns them not to or if the fully qualified name is not
displayed in the cross reference. The `sympy.module.<tab>` in IPython
will show all variables, including module names, that do not start with
an underscore, so people will import from anywhere. I have yet to come
across a library that shifts module layout without deprecation and
warning to users.