Improve `gf_edf_zassenhaus()`.
For the case p == 2, we use Algorithm 3.4.8 of [Cohen93], instead of the current procedure.

The current algorithm was failing to terminate on at least one known case (factoring cyclotomic_poly(17) mod 2).

A simple bugfix would have been to change the iteration to `for i in range(n - 1):`
when computing the polynomial `h` (`Tr` in Geddes), but Alg 3.4.8 is thought to
be better in practice.