orient_axis() relies on angles being real to simplify
orient_axis() uses normalize() (and thus magnitude()) internally for
establishing unit vectors to get the correct direction cosine matrices.
For space fixed successive orientations this means that Abs() terms get
wrapped into the results (because the orientation axes are all in the
first reference frame). Making the angles real is required for proper
simplification in this example. This may cause other unexpected behavior
to user code, but it is the correct mathematical answer now.