fix(core): make core assumptions fully deterministic
The core assumptions system could be non-deterministic in some cases due
to the storing a key value of None in the _assumptions dict while
handling an assumptions query. For example in a query like e.is_zero the
assumptions system would:

1. Store None in self._assumptions['zero']
2. Look for and evaluate _eval_is_zero
3. If _eval_is_zero then other facts (is_integer, is_real, etc) are
checked.
4. While checking the other facts zero=None is still stored in the
_assumptions dict so any code that checks e.is_zero receives a None
value.
5. Possibly later the value of is_zero becomes known e.g. if is_real
turns out to be False than is_zero must be False.

The problem here is that in step 4 is_zero gives None even though a
fully resolved is_zero query should give e.g. False. Since the different
facts attempted in step 3 are tried in a random order this can lead to
non-deterministic behaviour.

The reason None is stored in self._assumptions['zero'] is that without
then a query like e.is_zero that occurs recursively while trying to
evaluate e.is_zero leads to infinite recursion. The None is stored there
to prevent the infinite recursion but that is really just masking bugs
and replacing the infinite recursion with non-deterministic behaviour.

With the change in this commit it is now necessary that no assumptions
query ever recursively queries an assumption on self. Mostly there is no
reason for assumptions queries to do that and in many cases the change
here just deletes a redundant _eval_is_* method that isn't actually
needed because the resolver in the assumptions system can figure out the
result itself.

The other situation where this could be problematic is where an
assumptions query calls higher-level code that should not really be used
within the assumptions system because it makes things circular. The
higher-level code should be free to query assumptions on e and so should
not be called as part of an assumptions query on e.