Use better contrast colors for links
This removes the separate colors for visited and unvisited. We now use the
darker of the two in light mode and the ligher of the two in dark mode. These
colors satisfy AA color contrast, both against the background and against the
surrounding black/white text (see
https://github.com/sympy/sympy/pull/23159/#issuecomment-1078480215).

This replaces the "brand-content" color in light mode since this is primarily
used for links. We still use the lighter color for the API function names in
light mode, so that they pop out more (dark mode already has a lighter color
for API function names).