Use custom pygments styles with better color contrast in the docs
We use the same styles as before (sphinx for light mode and the Furo default
of native for dark mode), but they have been updated so that any text colors
that were lower than 4.5 color contrast have been modified. The styles mostly
look the same but some parts may be easier to read. The 4.5 color contrast
makes the docs WCAG AA compliant.