Proof of concept: Make Function classes behave as Basic instances This is a proof of concept to demonstrate that it is possible to make `Function` classes behave as if they are `Basic` instances, without interfering with instances of those `Function` classes. I'm currently looking for counter examples where this fails.
Some ideas to think about if something along these lines were to
be added to SymPy:

Should `isinstance(sin, Basic) == True`? It's possible to do this
technically, but it would potentially alter many code paths.

Should the information of "specific" functions (e.g. `sin`) be
stored in the `FunctionSymbol`s? This would allow a move towards
something along the lines of `Function(FunctionSymbol(), x, y, ...)`
by seperating changes in the library and the actual deprecation
process. Perhaps even allowing the current function system as well
as a new one to work at the same time.