Tone down the brand-primary color, and use it for the sidebar hover
It's only used in two places, the left sidebar hover, and the right sidebar
active page. This color was taken from the SymPy logo, and modified slightly
to have a 4.5 contrast ratio. Unlike the previous color, this one now has good
contrast, except in dark mode, where we use the old color for the black
background.