maint: ignore last commit for authors check in CI
When the mailmap_check.py script runs in CI it runs in a checkout of
temporary merge commit. The merge commit created by GitHub usually does
not use the same metadata as commits created locally by the author. This
required redundant additions to .mailmap. This commit makes it so that
when the script runs in CI the most recent commit is ignored.