Make `CRootOf` lazy on irreducibles.
If `f` is irreducible, and you construct `CRootOf(f, n)`, then the index `n` need not be adjusted.
This means we are free to postpone calculation of the isolating intervals until they are needed,
and this commit makes this so.

If `f` is composite, then the index `n` may need to be adjusted, which requires calculation of the
isolating intervals at construction time, so we cannot postpone it.