Ensure inert primes have `0` second generator.
We only need to fix this in the "easy case," since it is already correctly
implemented in the "hard case."

In the easy case, `T` stays irreducible over `FF(p)`, so if we lift back to `ZZ`, and
reduce mod `T` (as we were doing), we can get a non-zero poly, hence the bug.

In the hard case, the `_prime_decomp_maximal_ideal()` function passes an empty
list of generators to `_two_elt_rep()`, which ensures `alpha` is `0`.