fix(polys): fix pickling of poly objects for Python 3.11
This is needed after the changes in

  https://github.com/python/cpython/issues/70766

Since object now has a __getstate__ method we need to make sure not to
call it with any arguments. Probably the __getstate__ methods can be
simplified (or removed?) in light of the cpython changes but for now
this is a quick fix to restore previous behaviour.