test: fix use of Float/Rational in tests
This commit changes tests that assert comparisons between Floats and
Rationals as well as Floats with different precision. Where a tested
function returns a Float (or an expression involving Floats) then the
test should compare the returned value with a Float. Likewise if exact
Rationals are returned then the test should assert that.

There were also many tests that use Floats with specific precision but
compare against Floats with different precision. This commit ensures
that a correct precision Float is always used in these comparisons.