fix(printing): use PYFUNCTYPE to fix segfault in LLVM
In Python 3.10 the ctype.PYFUNCTYPE convention needs to be used for a
function that internally uses PyFloat_FromDouble as is done in the
LLVMJIT code printer in some cases. This commit changes the function to
PYFUNCTYPE whenever the return type is ctypes.py_object i.e. if the
jitted function returns a PyFloat or a PyTuple rather than a double.