Issue a warning when is_assumption is defined as a @property method
For example

>>> class Bad(Function):
...     @property
...     def is_zero(self):
...         return self.args[0].is_zero
./sympy/core/function.py:197: UserWarning: is_zero is defined as a property method on Bad. Doing so will break the automatic deduction of other assumptions. Either define is_zero as a simple bool or define _eval_is_zero instead.
  super().__init__(*args, **kwargs)

A few cases of this in library code have been fixed.

However, there are still a few instances of the warning when importing sympy

./sympy/polys/polytools.py:102: UserWarning: is_zero is defined as a property method on <class 'sympy.polys.polytools.Poly'>. Doing so will break the automatic deduction of other assumptions. Either define is_zero as a simple bool or define _eval_is_zero instead.
  class Poly(Basic):
./sympy/geometry/point.py:42: UserWarning: is_nonzero is defined as a property method on <class 'sympy.geometry.point.Point'>. Doing so will break the automatic deduction of other assumptions. Either define is_nonzero as a simple bool or define _eval_is_nonzero instead.
  class Point(GeometryEntity):
./sympy/geometry/point.py:42: UserWarning: is_zero is defined as a property method on <class 'sympy.geometry.point.Point'>. Doing so will break the automatic deduction of other assumptions. Either define is_zero as a simple bool or define _eval_is_zero instead.
  class Point(GeometryEntity):

(I haven't yet checked the parts of SymPy that aren't imported automatically).

These are all instances where is_zero is used as its own thing, and doesn't
really seem to be indented to be part of the assumptions system. I'm not sure
what should be done about them.

I haven't yet checked if this check affects the performance in any way. The
main thing to check would be the import time, since this runs when classes are
defined.

This is based on the discussion at
https://github.com/sympy/sympy/pull/23488#discussion_r900465301.