fix(printing): restore kwarg for Theano
This was changed because Aesara gives a warning when the broadcastable
argument is used and suggests changing it to shape. However shape is
incompatible with Theano so this commit restores the use of the
broadcastable argument that works with both Theano and Aesara although
it gives a warning under Aesara.