solve_undetermined_coeffs enforces single solution
Since flags are passed to solve, it is possible to get many different forms from this function. That has been changed and the assumption of a single solution is now enforced: the return value will be `[]` if there is no solution, a dictionary if there is a single solution, else None.