mailmap documents python2
I believe this to be useful for at least two reasons.

Current documentation either state to use `python bin/mailmap_check.py` or just
`bin/mailmap_check.py` which itself calles `python`. In both case, on ubuntu
this will run python2.

Worse, instead of printing "This script requires Python 3.8 or newer" it prints
an error message that python don't know what to do with the f-string. If the
f-string is removed, that it does not know pathlib. So, I ensure that the script
still work as before on 3.8 and prints a more useful message on python2.  I
admit I'm not fan of removing a f-string, however, since there was a single
f-string, I assume it was relatively acceptable.

I suppose most sympy contributor are at ease with those subtilities of
python2/3. However, this will at least be useful for people, like me, who only
wanted to contribute to improving documentation and not have to deal with
complexity of system administration.