functions: Fix inconsistency of zeta(-n,a) with Apostol's Intro to ANT
When n is a non-negative integer Apostol's Introduction to Analytic
Number Theory (Theorem 12.13 in the 1995 edition) states that the
Hurwitz zeta function ζ(-n, a) may be reduced to a simpler expression
involving a Bernoulli polynomial:

           -B     (a)
             n + 1
ζ(-n, a) = ───────────
              n + 1

This is valid for any complex a, so for example ζ(-1, -4) = -121/12,
agreeing with mpmath. However, SymPy and (catastrophically) Mathematica
return 119/12 for these arguments. This commit introduces the above
relation to SymPy, simplifying a small piece of code, as well as
correcting two test cases.

With this change, ζ(s, a) has only one branch cut along the negative
real axis for fixed noninteger s instead of infinitely many, as can be
seen by plotting in Mathematica.