functions: Refactor Bernoulli function evaluation
In the previous version of bernoulli()'s eval() the check for sym being
None appeared several times in the code, even though it is the most
salient feature differentiating between the Bernoulli numbers and
polynomials. This refactor puts that check near the root of the decision
tree and incidentally also allows suppressing explicitly provided sym=1,
renamed to x=1.