functions: Refactor zeta function evaluation
Luschny's introduction to the Bernoulli function, when it introduces the
generalised Bernoulli function as the analogue of the Hurwitz zeta
function, notes that the extra parameter "can be any complex number that
is not a nonpositive integer". The implication is that ζ(s,a) is
undefined if a is a nonpositive integer, *except* when s is also a
nonpositive integer, the latter allowing us to apply Apostol's identity.

On these grounds the decision tree for evaluating zeta() has been
refactored, removing the vestiges of two-argument Zeta[] behaviour and
making things like zeta(2,-2) strictly undefined. This new version also
suppresses explicit a=1 arguments like bernoulli().