functions: Properly extend harmonic() to complex arguments
The rewrite rule from harmonic() to polygamma() is – as hinted by the
presence of (-1)^m in the expression – only valid for positive integer
m. The generally accepted extension of harmonic() to complex arguments
is

H_{n,m} = ζ(m) - ζ(m, n+1)

as long as m ≠ 1. But 1 is a positive integer, so harmonic().evalf() is
still coded to call the polygamma() rewrite if m = 1.