functions: Restrict automatic zeta evaluation
In order to prevent potential infinite recursion when harmonic() is
extended to complex arguments, the scope of zeta(s,a)'s automatic
evaluation when a is an integer has to be restricted. Here the rewrite
is performed if and only if s is an Integer and a is a positive Integer,
ensuring that the harmonic part collapses into a Rational.