functions: Generalised Bernoulli function
In Peter Luschny's "An introduction to the Bernoulli function"
(https://arxiv.org/abs/2009.06743, will just be called Luschny in future
commits) a generalised (two-argument) Bernoulli function is defined. One
of the given representations is

          ⎧      1         for s = 0
B(s, a) = ⎨
          ⎩-s⋅ζ(1 - s, a)  otherwise

When s is a nonnegative integer this function reproduces the Bernoulli
polynomials; when in addition a = 1 (the default choice for the Hurwitz
zeta function's second argument, yielding Riemann's zeta) the Bernoulli
numbers are obtained. Hence B(1) = +1/2, which is how Jakob Bernoulli
originally defined his numbers in Ars Conjectandi (1713); this choice
gives the simpler form/wider range of validity than B(1) = -1/2 for very
many equations like Faulhaber's formula and the Euler–Maclaurin
summation formula.

We use the zeta representation of the Bernoulli function, despite
Luschny striving to avoid such dependence by defining it as an integral,
since mpmath provides a correct implementation of the Hurwitz zeta
function; SymPy's zeta() was also aligned to Hurwitz zeta in #23973.