functions: Generalised Genocchi function
Luschny defines the generalised Genocchi function in terms of the
generalised Bernoulli function (implemented in #23984) as

G(s, a) = 2^s (B(s, a/2) - B(s, (a+1) / 2))

which can be rather easily rewritten as

G(s, a) = 2 (B(s, a) - 2^s B(s, (a+1) / 2))

which is the form used here. When s is a nonnegative integer this
function reproduces the (so-called) Genocchi polynomials, which have
integer coefficients; when in addition a = 1 the Genocchi numbers are
obtained, with G(1) = -1.

The second term in the rewritten expression (2^s B(s, (a+1)/2)) is what
Luschny calls the central Bernoulli function, which reduces to the
central Bernoulli numbers when a = 0.