functions: Generalised Euler function
Luschny defines the generalised Euler function in terms of the
generalised Genocchi function (implemented in #23989) as

          ⎧   ⎛a⎞    ⎛a + 1⎞
          ⎪ -ψ⎜─⎟ + ψ⎜─────⎟  s = -1
          ⎪   ⎝2⎠    ⎝  2  ⎠
E(s, a) = ⎨
          ⎪   -G(s + 1, a)
          ⎪   ─────────────   s ≠ -1
          ⎩       s + 1

where the top expression is the limiting value of the bottom. When s is
a nonnegative integer the function reproduces the Euler polynomials, but
unlike the Bernoulli and Genocchi polynomials, a must be set to ½ *and*
the result must be multiplied by 2^s to obtain the Euler numbers. For
even more clarity Luschny disambiguates the latter with the designation
"secant", calling 2^s E(s, 1) the Euler tangent numbers.