functions: Espinosa–Moll generalised polygamma function
Olivier Espinosa and Victor Hugo Moll in "A Generalized Polygamma
Function" (http://www.math.tulane.edu/~vhm/papers_html/genoff.pdf)
define the titular function as

          ζ'(s+1, z) + (ψ(-s) + γ) ζ(s+1, z)
ψ(s, z) = ──────────────────────────────────
                        Γ(-s)

which is defined iff z is not a nonpositive integer. Either directly or
in the limit this expression reproduces the classical polygamma
functions when s is a nonnegative integer AND for any complex s always
satisfies the fundamental identity

∂
──(ψ(s, z)) = ψ(s+1, z)
∂z

but the definition implies ψ(-1, z) = logΓ(z) - log(2π)/2,
not just logΓ(z), which is fine.

Mathematica's PolyGamma[] also accepts complex arguments but does not
produce the same results as the above function, instead implementing a
generalisation by Grossman (1976). mpmath does not include the
generalisation, but Fredrik Johansson's other library Arb does:
https://arblib.org/acb.html?highlight=polygamma#c.acb_polygamma