Probabilist's Hermite polynomials
The plain or physicist's Hermite polynomials have leading coefficient
2^n, which leads to orthogonality with respect to the simplest possible
form of the weight function exp(-x^2) and is the specific normalisation
appearing in the solution to the Schrödinger equation for the quantum
harmonic oscillator, but leads to unnecessary complications everywhere
else.

Removing the twos in the 3-term recurrence relation leads to the monic
probabilist's version; its weight function of exp(-x^2/2) becomes the
standard normal distribution after normalising. This version also forms
the sign-alternated matching polynomial for the complete graph, a highly
interesting connection to combinatorics.