Cleanup documentation in preparation for sphinx-doc/sphinx#10807
This upstream Sphinx PR will make each function, class, and method docstring
appear in the table of contents (i.e., in the right sidebar). This means that
headers like

    sympify
    ^^^^^^^
    .. autofunction:: sympify

are now redundant.

This can be merged before the upstream Sphinx PR is merged, however, until a
Sphinx release is made with it, those pages will not have headers for the
functions. However, the use of the above one-header-per-function style was
done very inconsistently throughout the reference pages.

This PR also

- Changes all index.rst pages to use :titlesonly: instead of :maxdepth:. This
  makes the table of contents pages only show the table of contents for actual
  pages. With each function included in the ToC, showing them all on the index
  page made the table of contents quite large in many cases.

  These index pages are still mostly useless, especially given the left
  sidebar. I leave it to a future PR to either do something better with them
  or remove them entirely. This is related to #23333.

- I have split the mechancs, vector, and physics vector API toc into
  separate index pages. This does not affect the URLs of the pages themselves
  (the API pages were already in separate api/ subdirectories), but it does
  make the left sidebar put these API pages under a separate foldable API
  heading.

- A few very minor cleanups. I have left most things intact they way they
  were, even in instances where things could clearly be improved.

Fixes #23332.