Enable development docs deployment from GitHub Actions
The Travis builds no longer seem to be running, so the previous system with
doctr is no longer working.

I have already set up an SSH deploy key and added the keypair to the
respective repos.

I am making this PR against a branch in the main repo so that I can test it
before merging it.