maint(ci): skip archlinux SAGE job in GA
For now the archlinux job fails too often for reasons unrelated to
SAGE's use of SymPy so we skip Arch and only test Ubuntu. The
ubuntu-focal job also fails sometimes due to timeouts or network
problems but it mostly succeeds and does correctly pick up where a
change in SymPy affects SAGE. Mostly this is where there are printing
changes in SymPy because SAGE does most of its testing in the form of
doctests.