Add a flag nonsymbolic_exponent to Pow.rewrite(exp)
Setting it to False (it is True by default) disables rewriting cases that
create unevaluated Pow/exp objects.

This is used in solve to avoid a recursion error (fixes #24368) which is
caused by the call to rewrite(exp) in solve, which then gets rewritten back to
a Pow by Poly. There doesn't appear to be any benefit to solving to rewrite
powers with nonsymbolic exponents to exp.