Don't do rewrite(exp) in solve()
This can lead to recursion errors (e.g., #24368) because it is too easy for
the expression to get rewritten back to the way it was again in a recursive
call.

This still has some test failures. We need to handle trigonometric equations
in some other way (e.g., by doing a more targeted rewrite of them).