polys: remove the ring cache
Rings and fields like QQ[x] were cached in a cache that grows without
bound. This commit removes the cache altogether and changes places that
depend on the cache. The main need for the cache is that each ring
dynamically creates a class for its elements and then uses isinstance to
check if an object is an element of the ring. Here we change all places
that use isinstance to compare the rings instead since the rings already
have __eq__ defined.

A subsequent commit will restore some sort of ring cache but one that
does not grow without bound.