polys: cache the PolyRing.clone method
The PolyRing.clone method is used to generate rings similar to an existing
polynomial ring e.g. by PlyRing.drop. Since this ends up being used in the
inner loop of various operations like heugcd we cache the result of the clone
method so that as long as this ring is in memory it can keep references to any
related cloned rings without needing to rebuild them from scratch. This could
be an alternative to the global ring cache that was previously removed.