polys: cache MonomialOps codegen routines
Caches the most expensive part of constructing a polynomial ring which is
generating and compiling the code for monomial operations in the MonomialOps
class. First we have a LRU cache for MonomialOps instances of a given number
of generators. Then we also cache the compiled methods generated by each
MonomialOps instance.