feat(core): add equal_valued function
This function is equivalent to == i.e equal_valued(a, b) should be the
same as a == b for any SymPy expressions. However if it should ever be
decided that Float should compare uneqal to Rational then this function
could be used to replicate the previous behaviour.

See https://github.com/sympy/sympy/pull/20033