Replace the ManagedProperties metaclass with a base class with __init_subclass__
This is primarily a proof-of-concept for now. If we decide that we want to do
this, we should apply some cleanups to this. For example:

- ManagedProperties is still used in some other places in the codebase.
- We can probably replace _BasicBase with just moving some of the
  __init_subclass__ logic into Basic.__init__

I'm curious if this has any impact on performance, but it's hard to tell since
we still have the BasicMeta metaclass (with the ordering of classes), which is
harder to remove.