Removed unused variables
Variable C is an important parameter that determines the elliptic curve,
but is not required for the calculation.
It is also likely that the variable name C was used from [1].
In [1], the elliptic curve is defined as y^2=x^3+Cx^2+Ax+B.
Since this docstring uses a, it would be better to unify it with a.