fix(dsolve): handle simple DAE systems in dsolve
Fixes gh-24841

A mixture of algebraic and differential equations is known as a system
of differential algebraic equations. The machinery in dsolve_system can
handle simple cases of this but needs to take care not to misconstrue
them as a system of ODEs.

This commit changes it so that:

- _classify_linear_system no longer incorrectly identifies a system
  containing algebraic equations as being first order. On its own this
  prevents a wrong result for DAEs but would mean that dsolve cannot
  solve them.

- Teaches dsolve_system's strong component solver to recognise algebraic
  equations that are already in solution form like x(t) = sin(t) etc.
  Nothing special is needed to solve these but the solver needs to check
  for that case before trying to classify as a system of ODEs.

- Fixes a few other places that assume that all the functions to be
  solved for will appear as derivatives in the equations.

Note: The case of inconsistent initial conditions is not correctly
handled yet due to buggy code in solve_ics.