feat(dsolve): handle underdetermined ODE systems
This commit adds tests for cases of using dsolve with a mixture of ODEs
and algebraic equations. A number of changes are made to be able to
handle underdetermined cases such as

   x' = y'

Here attempting to solve algebraically for x' and y' will not give a unique
solution. We introduce the idea of parametric functions so that solving
here can give a system of ODEs like:

   x' = C1(t)
   y' = C1(t)

Then a general solution can be expressed in terms of the undefined
function C1. Various functions such as constant_renumber and
checksysodesol are modified to better handle cases such as this where
the number of equations does not necessarily equal the number of
unknowns.