feat(dsolve): improve dsolve_system for DAEs
Improve the way that dsolve_system handles mixed systems of algebraic
and differential equations. In some cases it is possible to solve for
lower order derivatives of a given function such as

  x' + y' = y
  x' + y' = 0

Here although this looks like a first order system for x and y the
equations actually impy that y = 0 and therefore that y' = 0 so this can
be reduced to

  x' = 0
  y  = 0

This commit adds a strategy based on first solving for all derivatives
of all variables algebraically. This gives an underdetermined system but
might reveal that some low order derivatives can be solve for
algebraically which can then be used to reduce the original system of
equations.