perf(core): reduce assumption checking in Mul(x,y)
A common case for Mul is a 2-arg Mul as this arises from any code using
e.g. a*b. The Mul.flatten function tries to special case this to
distribute a rational into an Add if either arg is a rational. This was
tested for in a suboptimal order as

    if not a.is_zero and a.is_Rational:

With large expressions a.is_zero can be expensive as it invokes the core
assumption system which causes many different assumption handles to run.
By constrast a.is_Rational is always a plain attribute look up giving
True or False so checking it first in order to shortcut checking for
a.is_zero obviously makes sense. Better than that though in the case
that a.is_Rational is True then a.is_zero is trivial so reversing the
order here eliminates all nontrivial is_zero assumption queries from
this line. Since every call to a*b goes through this line of code this
can give a significant speed boost for some macro benchmarks.