Use generalisation of Lucas' Theorem for binomial(n, m) mod q
Create a new function `binomial_mod` in the `residue_ntheory` module
which uses a generalisation of Lucas' Theorem to prime powers given
by Granville in
https://web.archive.org/web/20170202003812/http://www.dms.umontreal.ca/~andrew/PDF/BinCoeff.pdf
allowing the computation of binomial(n, m) mod p^q quickly when p is
sufficiently small.  This extends Lucas' Theorem to all integers n,
if we first factorise n, and then apply the Chinese Remainder Theorem
after applying the new method for each prime power.

Make the binomial class make use of this to calculate binomial(n, m)
modulo some q when q is less than the square root of n, and document
example usage of this in the binomial class docstring.  Also document
the interaction between `Mod` and `binomial` in the `Mod` class docstring.