Update mul.py
Change type annotation at line no 164 in sympy/sympy/core/mul.py

References to other Issues or PRs
Fixes #24871

Brief description of what is fixed or changed
"args: tTuple[Expr]" changed to "args: tTuple[Expr, ...] "

Other comments:
Also previously made a syntactical error.

The issue with the syntax tTuple[Expr] is that it does not allow for a tuple of variable length. The tTuple type annotation, if it is a user-defined type annotation, may have been defined to only allow tuples of a fixed length.

On the other hand, tuple[Expr, ...] is a built-in type annotation that allows for a tuple of any length, as indicated by the ellipsis .... This syntax is recommended for defining a tuple of variable length.

Release Notes