fix(polys): fix factor_list over EX domain
The factor_list function would fail over the EX domain when generators
were provided e.g.:

   factor_list((1 + sqrt(2))*(1 - sqrt(2)), x)

The failure comes from attempting to sort the factors when elements of
EX are not comparable. In this example actually both factors should be
part of the coeff rather than the factor list so that is fixed by
handling the case of zero degree polynomials in Poly.factor_list. It is
still possible to see the problem with ordering factors over EX though
with somethng like:

   factor_list((x + sqrt(2)*pi)*(x - sqrt(2)*pi), x)

The __lt__ method has been added to EX so that its elements can be
ordered as expected by factor_list.