Fix Polygon.angles returning conjugates instead of interior angles
Changes address issue #24885.

To find the angles of a polygon without making assumptions about it, you have
to take into account all of its vertices in some way. Consider that if you were
given a polygon with one vertex removed, it would be impossible to tell which
angles were the interior angles and which angles were the conjugates of the
interior angles.

Previously the code determined the orientation of the polygon by checking the
orientation of only one of the angles of the polygon. I believe the method used
might work if the polygon is assumed to be convex but definitely does not work
in all cases.

The new code uses the fact that the sum of the interior angles of any polygon is
always (n-2)*180. If the computed angles don’t sum up to this, the angles
function returns the conjugates of the wrong angels instead.

I also renamed _isright to _is_clockwise to avoid confusion and better reflect
what the function does.

A note on terminology: the conjugate of an angle x is 360 -x; conjugate angles
are pairs of angles that add up to 360 degrees.

Co-authored-by: smichr